fix Int32 unbox in Enqueue max-rank
MAX(rank) on an int4 column returns a boxed Int32; unboxing directly to long throws InvalidCastException. Cast to int.

+++ dashboard/src/Api/Queue/QueueStore.cs
@@ -64,7 +64,7 @@ public sealed class QueueStore : IQueueStore
 
         await using var rankCmd = new NpgsqlCommand(
             "SELECT COALESCE(MAX(rank), 0) FROM queue;", conn, tx);
-        var maxRank = (int)(long)(await rankCmd.ExecuteScalarAsync() ?? 0L);
+        var maxRank = (int)(await rankCmd.ExecuteScalarAsync() ?? 0);
 
         await using var insertCmd = new NpgsqlCommand(
             """
